Add common functions to Utils class
