+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/models/RadIssue.java
@@ -53,9 +53,7 @@ public class RadIssue {
     @JsonFormat(shape = JsonFormat.Shape.OBJECT)
     public enum State {
         OPEN("Open"),
-        CLOSED("Closed"),
-        MERGED("Merged"),
-        ARCHIVED("Archived");
+        CLOSED("Closed");
 
         public final String status;
